requirements.txt
# Core dependencies
pandas>=2.0.0
requests>=2.31.0
python-dateutil>=2.8.2
pyyaml>=6.0

# Korea Investment API
finance-datareader>=0.9.50
pykrx>=1.0.40

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Utilities
setuptools>=68.0.0
